Summary:  Fixes #2 - Final report is javascript, not JSON
**Problem:**
The final report is valid javascript, but not valid JSON.  This makes it
more difficult than necessary to suck it into another program and use
the data.

**Analysis:**
This is caused by using 'util.inspect' instead of JSON.stringify when
printing out the representation of the final report.

Updated to use the latter.

**Testing:**
Manual

**Documentation:**
No change - the documentation should be updated to describe the final
report as JSON, but a large README update is coming soon...