Summary:  Fixes #3 - support public key auth
**Problem:**
The tool doesn't support public key auth, which is a common auth case.

**Analysis:**
Public key auth is common enough to be a very useful feature for the
performance tool, and different enough in process under the covers from
password auth that it might conceivably have a performance impact.

Added an option to use a public key, and made the password also an
option, so that you choose which auth by which option you use.

**Testing:**
Manual testing

**Documentation:**
None, but it needs an update to the README, which is coming soon.