Summary:  Intended fix for #4 - support SSH agent forwarding
**Problem:**
The tool doesn't support SSH Agent forwarding, which is a useful feature for connection hopping.

**Analysis:**
Add an agent option to the CLI, and pass that through to the SSH2 library so that it will use the agent.

See Testing section, but the underlying library doesn't actually support this feature yet.

**Testing:**
Manual.  Unsuccessful.  Upon inspection, the SSH2 library doesn't actually support the client side of SSH agent forwarding.

**Documentation:**
Help and README text