Summary:  Fixes #7 - update copyright info
**Problem:**
Copyright info too restrictive for GPLv2 license.

**Analysis:**
Use the standard GPLv2 copyright header instead.

**Testing:**
None - when automated tests are created, they should check the licensing info.

**Documentation:**
This change IS documentation, in a way.  The README is unchanged, though, other than its own
copyright, as it does not address copyright of the code, in general.