Summary:  Fixes #5 - add SSH debug option
**Problem:**
No way to see what the lower-level SSH library is doing during SSH operations.

**Analysis:**
It is often useful when debugging to see some lower-level SSH data about both sides of a connection.
Not presenting that option limits the usefulness of this tool.  While it is not meant to be a diagnostic tool,
it is low-cost to add a little of this functionality by utilizing SSH2's 'debug' option.

SSH2 provides an argument which can be set to a callback function for debugging.  Any time SSH2 has some debug
output, it will call the provided function, if any.

Added a CLI option to enable debug printing, and set the SSH2 debug callback.  If the option is set, the callback
will print the SSH2 message to stderr.

**Testing:**
Manual

**Documentation:**
Help text.