Summary:  Fixes #9 - maximum js string size error
**Problem:**
When running with large files as the scp upload objects, the workers regularly crash with a
"maximum string size" error.

**Analysis:**
This is caused by leftover debug code from when SCP upload was being implemented.  The code recorded bytes
sent up to the remote side in a string.  If the transfer failed, the sent bytes were printed to the console
for inspection and debugging.

This code is no longer needed, and there was some similarly crusty code, much of it commented out.  That has
all been removed.

**Testing:**
Manually tested 1G SCP uploads, which were failing, and now work.

**Documentation:**
None, this was a bug which was not documented, and now does not need to be.