Summary:  Fixes #10 - scp upload backpressure
**Problem:**
The tool continues to buffer data AND count that data as 'sent' for the purposes of throughput stats,
even after the network stream says it's buffering and not sending.

This leads to incorrect, and sometimes impossible, numbers (like 4x the line rate).

**Analysis:**
Add backpressure code to the SCP-up implementation - this, as is, is a big problem.

**Testing:**
Manual

**Documentation:**
None - no current documentation addresses this.